requirements.txt
fastapi
uvicorn
pydantic
motor
google-auth
google-auth-oauthlib
google-api-python-client
python-dotenv
